refactor: organize `filter` files
